自機と追手を壁が遮蔽しているかチェックするメソッドを追加
